Composite and Tn3, optional arrays (#12)
* Add some files to run some more analyses for Ilya

These may or may not make it into the paper depending on what we find.
If it turns into something real I will rename things.

* Stream gene finder output directly

Previously I was decompressing the gene finder output and saving it to disk, but this was unnecessary and wasn't part of the pipeline. Now the operons are decompressed and streamed directly from the original files.

* Composite and Tn3 subsets

Add steps in the pipeline to break out the composite transposon and Tn3
subsets.
Refactor the pipeline script to make the additional steps clear and
consistent.

* Refactor scripts

I rewrote the main pipeline script since it was getting too confusing
for even the author to comprehend after adding in the Tn3/composite
code.

Data is now split into two categories: systems where CRISPR arrays were
mandatory, and systems where they were optional. All other requirements
are the same. Within those directories, we also now split out the data
based on whether it might contain Tn3 or composite transposon systems.
These are redundant subsets of the systems in the "all" directory, with
the intention that they are easier to browse manually.

* Finalize plotting code

The pipeline will now plot operons for manual curation and also plot all
systems that were re-BLASTed.