Add non-Tn7 data.
Adds a subset of the non-Tn7 data. I've included:
   - the results of re-BLASTing interesting systems
   - images of all systems (with annotations only from our curated
databases)
   - images of all re-BLASTed systems (with annotations from our curated
databases as well as the Swissprot/TrEMBL/tRNA databases)
   - the "fully analyzed" system data, which has systems without active
nucleases that passed all of our filters, and also has the inverted
repeat and self-targeting spacer analyses performed

These files, as well as the ones not included, can all be reproduced
deterministically if the original metagenomic database is available.